+++ app/src/main/java/de/alextape/androidcamera/CameraActivity.java
@@ -1,20 +1,11 @@
 package de.alextape.androidcamera;
 
 import android.app.Activity;
-import android.content.Context;
 import android.hardware.Camera;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.Display;
-import android.view.Surface;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
 import android.view.View;
-import android.view.WindowManager;
 import android.widget.FrameLayout;
-
-import java.io.IOException;
-import java.util.List;
 
 public class CameraActivity extends Activity {
 
@@ -91,159 +82,10 @@ public class CameraActivity extends Activity {
         }
         if (mPreview != null) {
             mPreview.destroyDrawingCache();
-            mPreview.mCamera = null;
+            mPreview.resetCamera();
         }
     }
 
-    class CameraPreview extends SurfaceView implements SurfaceHolder.Callback {
 
-        private SurfaceHolder mHolder;
-        private Camera mCamera;
-        private Context mContext;
-        private Camera.Size mPreviewSize;
-        private List<Camera.Size> mSupportedPreviewSizes;
-        private List<String> mSupportedFlashModes;
-        private View mCameraView;
-
-        public CameraPreview(Context context, Camera camera, View cameraView) {
-            super(context);
-
-            mCameraView = cameraView;
-            mContext = context;
-            setCamera(camera);
-
-            mHolder = getHolder();
-            mHolder.addCallback(this);
-            mHolder.setKeepScreenOn(true);
-            mHolder.setType(SurfaceHolder.SURFACE_TYPE_PUSH_BUFFERS);
-        }
-
-        public void startCameraPreview() {
-            try {
-                mCamera.setPreviewDisplay(mHolder);
-                mCamera.startPreview();
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-
-        private void setCamera(Camera camera) {
-            mCamera = camera;
-            mSupportedPreviewSizes = mCamera.getParameters().getSupportedPreviewSizes();
-            mSupportedFlashModes = mCamera.getParameters().getSupportedFlashModes();
-
-            if (mSupportedFlashModes != null && mSupportedFlashModes.contains(Camera.Parameters.FLASH_MODE_AUTO)) {
-                Camera.Parameters parameters = mCamera.getParameters();
-                parameters.setFlashMode(Camera.Parameters.FLASH_MODE_AUTO);
-                mCamera.setParameters(parameters);
-            }
-
-            requestLayout();
-        }
-
-        public void surfaceCreated(SurfaceHolder holder) {
-            try {
-                mCamera.setPreviewDisplay(holder);
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-        public void surfaceDestroyed(SurfaceHolder holder) {
-            if (mCamera != null) {
-                mCamera.stopPreview();
-            }
-        }
-
-        public void surfaceChanged(SurfaceHolder holder, int format, int w, int h) {
-            if (mHolder.getSurface() == null) {
-                return;
-            }
-            try {
-                Camera.Parameters parameters = mCamera.getParameters();
-
-                parameters.setFocusMode(Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE);
-                //parameters.setFocusMode(Camera.Parameters.FOCUS_MODE_CONTINUOUS_VIDEO;
-
-                if (mPreviewSize != null) {
-                    Camera.Size previewSize = mPreviewSize;
-                    parameters.setPreviewSize(previewSize.width, previewSize.height);
-                }
-
-                mCamera.setParameters(parameters);
-                mCamera.startPreview();
-
-            } catch (Exception e) {
-                e.printStackTrace();
-            }
-        }
-
-        protected void onMeasure(int widthMeasureSpec, int heightMeasureSpec) {
-
-            final int width = resolveSize(getSuggestedMinimumWidth(), widthMeasureSpec);
-            final int height = resolveSize(getSuggestedMinimumHeight(), heightMeasureSpec);
-            setMeasuredDimension(width, height);
-
-            if (mSupportedPreviewSizes != null) {
-                mPreviewSize = getOptimalPreviewSize(mSupportedPreviewSizes, width, height);
-            }
-        }
-
-        protected void onLayout(boolean changed, int left, int top, int right, int bottom) {
-            if (changed) {
-                final int width = right - left;
-                final int height = bottom - top;
-
-                int previewWidth = width;
-                int previewHeight = height;
-
-                if (mPreviewSize != null) {
-                    Display display = ((WindowManager) mContext.getSystemService(Context.WINDOW_SERVICE)).getDefaultDisplay();
-
-                    switch (display.getRotation()) {
-                        case Surface.ROTATION_0:
-                            previewWidth = mPreviewSize.height;
-                            previewHeight = mPreviewSize.width;
-                            mCamera.setDisplayOrientation(90);
-                            break;
-                        case Surface.ROTATION_90:
-                            previewWidth = mPreviewSize.width;
-                            previewHeight = mPreviewSize.height;
-                            break;
-                        case Surface.ROTATION_180:
-                            previewWidth = mPreviewSize.height;
-                            previewHeight = mPreviewSize.width;
-                            break;
-                        case Surface.ROTATION_270:
-                            previewWidth = mPreviewSize.width;
-                            previewHeight = mPreviewSize.height;
-                            mCamera.setDisplayOrientation(180);
-                            break;
-                    }
-                }
-
-                final int scaledChildHeight = previewHeight * width / previewWidth;
-                mCameraView.layout(0, height - scaledChildHeight, width, height);
-            }
-        }
-
-        private Camera.Size getOptimalPreviewSize(List<Camera.Size> sizes, int width, int height) {
-
-            Camera.Size optimalSize = null;
-
-            final double ASPECT_TOLERANCE = 0.1;
-            double targetRatio = (double) height / width;
-
-            for (Camera.Size size : sizes) {
-
-                if (size.height != width) continue;
-                double ratio = (double) size.width / size.height;
-                if (ratio <= targetRatio + ASPECT_TOLERANCE && ratio >= targetRatio - ASPECT_TOLERANCE) {
-                    optimalSize = size;
-                }
-            }
-            return optimalSize;
-        }
-    }
 
 }
